requirements.txt
torch>=2.7.0
transformers>=4.52.3
sentence-transformers>=4.1.0
faiss-cpu>=1.11.0
numpy>=2.2.6
PyPDF2>=3.0.1
python-docx>=1.1.2
async-timeout>=5.0.1
